Adding controller

+++ domian/src/main/java/com/eainde/ddd/domain/UserId.java
@@ -1,21 +1,18 @@
 package com.eainde.ddd.domain;
 
+import com.eainde.ddd.Wrapped;
+import com.google.common.base.Preconditions;
+
 import org.immutables.value.Value;
-
-import com.eainde.ddd.Wrapped;
-import com.google.common.base.Preconditions;
-
-import com.eainde.ddd.Wrapped;
-import com.google.common.base.Preconditions;
 
 @Value.Immutable
 @Wrapped
-public abstract class UserId extends Domain<Integer> {
+public abstract class UserId extends Domain<Long> {
   private static final String USER_ID_VALIDATION_FAILED_MESSAGE = "user id cannot be 0";
 
   @Value.Check
   public void check() {
-    final int value = value();
+    final long value = value();
     Preconditions.checkState(value != 0, USER_ID_VALIDATION_FAILED_MESSAGE);
   }
 }
